% Auto-generated: do not edit by hand
\name{arrayLayer}

\alias{arrayLayer}

\title{ArrayLayer component}

\description{
ArrayLayer component for react-sigplot/ArrayLayer.
}

\usage{
arrayLayer(id=NULL, data=NULL, layerOptions=NULL, options=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{List of numeric | list of numericss. 1D/2D Array of `Number` types to render in Arraylayer}

\item{layerOptions}{Named list. Options about the Arraylayer
@see See [sigplot.layer1d](https://github.com/LGSInnovations/sigplot/blob/master/js/sigplot.layer1d.js)
@see See [sigplot.layer2d](https://github.com/LGSInnovations/sigplot/blob/master/js/sigplot.layer2d.js)}

\item{options}{Named list. Header options for `data` for the Arraylayer}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

