% Auto-generated: do not edit by hand
\name{pipeLayer}

\alias{pipeLayer}

\title{PipeLayer component}

\description{
PipeLayer component for react-sigplot/PipeLayer.
}

\usage{
pipeLayer(id=NULL, data=NULL, layerOptions=NULL, options=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{List of numeric | list of numericss. 1D/2D Array of `Number` types to render in PipeLayer}

\item{layerOptions}{Named list. Options about the PipeLayer
@see See [sigplot.layer1d](https://github.com/LGSInnovations/sigplot/blob/master/js/sigplot.layer1d.js)
@see See [sigplot.layer2d](https://github.com/LGSInnovations/sigplot/blob/master/js/sigplot.layer2d.js)}

\item{options}{Named list. Header options for `data` for the PipeLayer}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

