% Auto-generated: do not edit by hand
\name{sigPlot}

\alias{sigPlot}

\title{SigPlot component}

\description{
SigPlot component for react-sigplot/Sigplot.
}

\usage{
sigPlot(children=NULL, id=NULL, display=NULL, height=NULL,
options=NULL, style=NULL, width=NULL)
}

\arguments{
\item{children}{Logical | numeric | character | named list | unnamed list | list of logical | numeric | character | named list | unnamed lists. react-sigplot *Layer elements contained within the SigPlot.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{display}{Character. CSS 'display' property for the SigPlot}

\item{height}{Numeric. Height of the SigPlot div}

\item{options}{Named list. SigPlot plot-level options
@see See [sigplot.Plot Docs](http://sigplot.lgsinnovations.com/html/doc/sigplot.Plot.html)}

\item{style}{Named list. Styles object for any other CSS styles on the SigPlot div}

\item{width}{Numeric. Width of the SigPlot div}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

